requirements.txt
Flask-RESTful==0.3.8
PyJWT===1.4.0
Flask-SQLAlchemy==2.4.1